// src/proto/types/users/User.ts
// Original file: src/proto/proto/user.proto

import type { Profile as _profiles_Profile, Profile__Output as _profiles_Profile__Output } from '../profiles/Profile';
import type { Company as _companies_Company, Company__Output as _companies_Company__Output } from '../companies/Company';
import type { Connection as _connections_Connection, Connection__Output as _connections_Connection__Output } from '../connections/Connection';
import type { IndividualRecord as _users_IndividualRecord, IndividualRecord__Output as _users_IndividualRecord__Output } from '../users/IndividualRecord';
import type { FundingRoundRecord as _users_FundingRoundRecord, FundingRoundRecord__Output as _users_FundingRoundRecord__Output } from '../users/FundingRoundRecord';
import type { OrganizationRecord as _users_OrganizationRecord, OrganizationRecord__Output as _users_OrganizationRecord__Output } from '../users/OrganizationRecord';

export interface User {
  'id'?: (number);
  'createdAt'?: (string);
  'updatedAt'?: (string);
  'createdBy'?: (string);
  'updatedBy'?: (string);
  'state'?: (string);
  'email'?: (string);
  'password'?: (string);
  'phone'?: (string);
  'role'?: (string);
  'profileType'?: (string);
  'phonePrefix'?: (string);
  'thumb'?: (string);
  'background'?: (string);
  'profileId'?: (number);
  'companyId'?: (number);
  'profile'?: (_profiles_Profile | null);
  'company'?: (_companies_Company | null);
  'sub'?: (string);
  'followingList'?: (_connections_Connection)[];
  'followerList'?: (_connections_Connection)[];
  'individualRecords'?: (_users_IndividualRecord)[];
  'fundingRoundRecords'?: (_users_FundingRoundRecord)[];
  'organizationRecords'?: (_users_OrganizationRecord)[];
  '_id'?: "id";
  '_createdAt'?: "createdAt";
  '_updatedAt'?: "updatedAt";
  '_createdBy'?: "createdBy";
  '_updatedBy'?: "updatedBy";
  '_state'?: "state";
  '_email'?: "email";
  '_password'?: "password";
  '_phone'?: "phone";
  '_role'?: "role";
  '_profileType'?: "profileType";
  '_phonePrefix'?: "phonePrefix";
  '_thumb'?: "thumb";
  '_background'?: "background";
  '_profileId'?: "profileId";
  '_companyId'?: "companyId";
  '_profile'?: "profile";
  '_company'?: "company";
  '_sub'?: "sub";
}

export interface User__Output {
  'id'?: (number);
  'createdAt'?: (string);
  'updatedAt'?: (string);
  'createdBy'?: (string);
  'updatedBy'?: (string);
  'state'?: (string);
  'email'?: (string);
  'password'?: (string);
  'phone'?: (string);
  'role'?: (string);
  'profileType'?: (string);
  'phonePrefix'?: (string);
  'thumb'?: (string);
  'background'?: (string);
  'profileId'?: (number);
  'companyId'?: (number);
  'profile'?: (_profiles_Profile__Output | null);
  'company'?: (_companies_Company__Output | null);
  'sub'?: (string);
  'followingList': (_connections_Connection__Output)[];
  'followerList': (_connections_Connection__Output)[];
  'individualRecords': (_users_IndividualRecord__Output)[];
  'fundingRoundRecords': (_users_FundingRoundRecord__Output)[];
  'organizationRecords': (_users_OrganizationRecord__Output)[];
  '_id': "id";
  '_createdAt': "createdAt";
  '_updatedAt': "updatedAt";
  '_createdBy': "createdBy";
  '_updatedBy': "updatedBy";
  '_state': "state";
  '_email': "email";
  '_password': "password";
  '_phone': "phone";
  '_role': "role";
  '_profileType': "profileType";
  '_phonePrefix': "phonePrefix";
  '_thumb': "thumb";
  '_background': "background";
  '_profileId': "profileId";
  '_companyId': "companyId";
  '_profile': "profile";
  '_company': "company";
  '_sub': "sub";
}