// src/types/apiErrors.ts
export type TApiError =
  | 'ERROR_PHONE_EMAIL_ALREADY_EXIST'
  | 'ERROR_INCORRECT_USERNAME_PASSWORD'
  | 'ERROR_INCORRECT_PASSWORD'
  | 'ERROR_INVALID_TOKEN'
  | 'ERROR_UNAUTHORIZED'
  | 'ERROR_NOT_FOUND'
  | 'ERROR_USER_ALREADY_ACTIVATED'
  | 'ERROR_INVALID_VERIFICATION_TOKEN'
  | 'ERROR_FORBIDDEN'
  | 'ERROR_POST_SLUG_EXIST'
  | 'ERROR_SYSTEM_ERROR'
  | 'ERROR_USER_ALREADY_EXIST'
  | 'ERROR_USER_CONNECTED'
  | 'ERROR_BAD_REQUEST';

export enum ApiErrorEnum {
  PHONE_EMAIL_ALREADY_EXIST = 'ERROR_PHONE_EMAIL_ALREADY_EXIST',
  INCORRECT_USERNAME_PASSWORD = 'ERROR_INCORRECT_USERNAME_PASSWORD',
  INCORRECT_PASSWORD = 'ERROR_INCORRECT_PASSWORD',
  INVALID_TOKEN = 'ERROR_INVALID_TOKEN',
  UNAUTHORIZED = 'ERROR_UNAUTHORIZED',
  NOT_FOUND = 'ERROR_NOT_FOUND',
  USER_ALREADY_ACTIVATED = 'ERROR_USER_ALREADY_ACTIVATED',
  INVALID_VERIFICATION_TOKEN = 'ERROR_INVALID_VERIFICATION_TOKEN',
  FORBIDDEN = 'ERROR_FORBIDDEN',
  POST_SLUG_EXIST = 'ERROR_POST_SLUG_EXIST',
  SYSTEM_ERROR = 'ERROR_SYSTEM_ERROR',
  USER_ALREADY_EXIST = 'ERROR_USER_ALREADY_EXIST',
  USER_CONNECTED = 'ERROR_USER_CONNECTED',
  BAD_REQUEST = 'BAD_REQUEST',
}
export enum ApiErrorCode {
  BAD_REQUEST = 400,
  NOT_FOUND = 404,
  OK = 200,
  CREATED = 201,
}